Refactor GIC initialization code
