+++ examples/mps3-an536/src/lib.rs
@@ -97,3 +97,42 @@ impl InterruptHandler {
         (self.function)(self.int_id);
     }
 }
+
+#[cfg(feature = "gic")]
+/// Offset from PERIPHBASE for GIC Distributor
+pub const GICD_BASE_OFFSET: usize = 0x0000_0000;
+
+#[cfg(feature = "gic")]
+/// Offset from PERIPHBASE for the first GIC Redistributor
+pub const GICR_BASE_OFFSET: usize = 0x0010_0000;
+
+#[cfg(feature = "gic")]
+/// Initialize the GICv3 controller for a single core
+///
+/// Returns the initialized GicV3 instance for further configuration
+pub fn init_gic() -> arm_gic::gicv3::GicV3<'static> {
+    use arm_gic::{gicv3::GicV3, UniqueMmioPointer};
+    use core::ptr::NonNull;
+    use semihosting::println;
+
+    // Get the GIC address by reading CBAR
+    let periphbase = cortex_ar::register::ImpCbar::read().periphbase();
+    println!("Found PERIPHBASE {:010p}", periphbase);
+    let gicd_base = periphbase.wrapping_byte_add(GICD_BASE_OFFSET);
+    let gicr_base = periphbase.wrapping_byte_add(GICR_BASE_OFFSET);
+
+    // Initialise the GIC.
+    println!(
+        "Creating GIC driver @ {:010p} / {:010p}",
+        gicd_base, gicr_base
+    );
+    let gicd = unsafe { UniqueMmioPointer::new(NonNull::new(gicd_base.cast()).unwrap()) };
+    let gicr = NonNull::new(gicr_base.cast()).unwrap();
+    let mut gic = unsafe { GicV3::new(gicd, gicr, 1, false) };
+
+    println!("Calling git.setup(0)");
+    gic.setup(0);
+    arm_gic::gicv3::GicCpuInterface::set_priority_mask(0x80);
+
+    gic
+}
